feat:Split BodyInput into TextBodyInput and FormBodyInput.
