+++ src/ui/components/common/mod.rs
@@ -1,3 +1,4 @@
 // src/ui/components/common/mod.rs
 pub mod dropdown;
 pub mod edit_context_menu;
+pub mod scrollbar;
